requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
pydantic==2.5.2
openai==1.3.5
python-dotenv==1.0.0
yt-dlp==2023.11.16
azure-cognitiveservices-speech==1.31.0
pymongo==4.6.1
python-jose[cryptography]==3.3.0
requests==2.31.0
google-api-python-client==2.108.0 